refactor : "이미지 패키지 변경"
