+++ src/main/java/challenging/application/global/images/S3PresignedImageService.java
@@ -1,4 +1,4 @@
-package challenging.application.images;
+package challenging.application.global.images;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
